pulled methods implementation into the abstract class, little tweaks for proper display
