+++ src/main/java/tkcy/simpleaddon/common/metatileentities/multiprimitive/MetaTileEntityMultiblockTank.java
@@ -140,18 +140,5 @@ public class MetaTileEntityMultiblockTank extends MetaTileEntityMultiblockStorag
         return this.fluidInventory.drain(Integer.MAX_VALUE, false);
     }
 
-    @Override
-    public String getPercentageTranslationKey() {
-        return "tkcysa.multiblock.modulable_storage.fill.percentage";
-    }
 
-    @Override
-    public String getCapacityTranslationKey() {
-        return "tkcysa.multiblock.modulable_storage.capacity";
-    }
-
-    @Override
-    public String getContentTextTranslationKey() {
-        return "tkcysa.multiblock.modulable_storage.content";
-    }
 }
